+++ src/main/java/blackjack/domain/card/exception/ScoreException.java
@@ -1,4 +1,4 @@
-package blackjack.card.exception;
+package blackjack.domain.card.exception;
 
 public class ScoreException extends RuntimeException {
     public ScoreException(String message) {
