+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -1,6 +1,6 @@
-package blackjack.user;
+package blackjack.domain.user;
 
-import blackjack.card.Score;
+import blackjack.domain.card.Score;
 
 public class Dealer extends AbstractUser {
     private static final String NAME = "딜러";
