+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack.card;
+package blackjack.domain.card;
 
 import org.junit.jupiter.api.Test;
