+++ src/test/java/blackjack/domain/user/AbstractUserTest.java
@@ -1,9 +1,9 @@
-package blackjack.user;
+package blackjack.domain.user;
 
-import blackjack.card.Card;
-import blackjack.card.Score;
-import blackjack.card.Symbol;
-import blackjack.card.Type;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Score;
+import blackjack.domain.card.Symbol;
+import blackjack.domain.card.Type;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
